+++ sdk/storage/azure-storage-files-shares/inc/azure/storage/files/shares/share_responses.hpp
@@ -116,7 +116,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares { names
     std::string ETag;
     std::string LastModified;
     int64_t ContentLength = 0;
-    FileShareHttpHeaders HttpHeaders;
+    ShareFileHttpHeaders HttpHeaders;
     Storage::Metadata Metadata;
     Azure::Core::Nullable<bool> IsServerEncrypted;
   };