+++ sdk/storage/azure-storage-files-shares/test/share_directory_client_test.cpp
@@ -26,11 +26,11 @@ namespace Azure { namespace Storage { namespace Test {
 
   void FileShareDirectoryClientTest::TearDownTestSuite() { m_shareClient->Delete(); }
 
-  Files::Shares::Models::FileShareHttpHeaders
+  Files::Shares::Models::ShareFileHttpHeaders
   FileShareDirectoryClientTest::GetInterestingHttpHeaders()
   {
-    static Files::Shares::Models::FileShareHttpHeaders result = []() {
-      Files::Shares::Models::FileShareHttpHeaders ret;
+    static Files::Shares::Models::ShareFileHttpHeaders result = []() {
+      Files::Shares::Models::ShareFileHttpHeaders ret;
       ret.CacheControl = std::string("no-cache");
       ret.ContentDisposition = std::string("attachment");
       ret.ContentEncoding = std::string("deflate");